+++ gtfstk/plotter.py
@@ -1,10 +1,21 @@
 """
 Functions for plotting various graphs related to GTFS feeds.
-Experimental.
+Optional.
+Requires Matplotlib.
 """
-
 import pandas as pd
-import matplotlib.pyplot as plt
+import importlib
+
+
+# Import matplotlib if it is installed
+loader = importlib.find_loader('matplotlib')
+if loader is None:
+    print("Warning: matplotlib is not installed, "\
+      "so the plotting functions (those in module gtfstk.plotter) "\
+      "will not work.")
+else:
+    import matplotlib.pyplot as plt
+
 
 def plot_headways(stats, max_headway_limit=60):
     """
@@ -56,13 +67,14 @@ def plot_headways(stats, max_headway_limit=60):
     xlabels = ['minutes', 'minutes']
     fig, axes = plt.subplots(nrows=1, ncols=2)
     for (i, f) in enumerate(data_frames):
-        f.plot(kind='barh', ax=axes[i], figsize=(10, max(n/9, 10)))
+        f.plot(kind='barh', ax=axes[i], figsize=(10, max(n/9, 10)),
+          color=['steelblue', 'darkslateblue'])
         axes[i].set_title(titles[i])
         axes[i].set_xlabel(xlabels[i])
         axes[i].set_ylabel(ylabels[i])
     return fig
 
-def plot_routes_time_series(routes_time_series):
+def plot_feed_time_series(feed_time_series):
     """
     Given a routes time series data frame,
     sum each time series indicator over all routes, 
@@ -74,19 +86,12 @@ def plot_routes_time_series(routes_time_series):
     Take the resulting figure ``f`` and do ``f.tight_layout()``
     for a nice-looking plot.
     """
-    rts = routes_time_series
-    if 'route_id' not in rts.columns.names:
-        return
-
-    # Aggregate time series
-    f = compute_feed_time_series(rts)
+    fts = feed_time_series.copy()
 
     # Reformat time periods
-    f.index = [t.time().strftime('%H:%M') 
-      for t in rts.index.to_datetime()]
+    fts.index = [t.time().strftime('%H:%M') 
+      for t in fts.index.to_datetime()]
     
-    #split_directions = 'direction_id' in rts.columns.names
-
     # Split time series by into its component time series by indicator type
     # stats = rts.columns.levels[0].tolist()
     stats = [
@@ -96,7 +101,7 @@ def plot_routes_time_series(routes_time_series):
       'service_duration',
       'service_speed',
       ]
-    ts_dict = {stat: f[stat] for stat in stats}
+    ts_dict = {stat: fts[stat] for stat in stats}
 
     # Create plots  
     pd.options.display.mpl_style = 'default'
@@ -110,7 +115,8 @@ def plot_routes_time_series(routes_time_series):
         else:
             stacked = True
         ts_dict[stat].plot(ax=axes[i], alpha=alpha, 
-          kind='bar', figsize=(8, 10), stacked=stacked, width=1)
+          kind='bar', figsize=(8, 15), stacked=stacked, width=1,
+          color='steelblue')
         axes[i].set_title(titles[i])
         axes[i].set_ylabel(units[i])